+++ src/cli.js
@@ -147,13 +147,13 @@ Project type-specific commands:
     Serve a React app from ${opt('entry')}
 
     Options:
-      ${opt('--no-hmre')}  disable use of React Transform for Hot Module Replacement & Errors
+      ${opt('--no-hmr')}  disable use of Fast Refresh for Hot Module Replacement
 
   ${cmd('nwb serve-react-demo')}
     Serve a React demo app from demo/src/index.js.
 
     Options:
-      ${opt('--no-hmre')}  disable use of React Transform for Hot Module Replacement & Errors
+      ${opt('--no-hmr')}  disable use of Fast Refresh for Hot Module Replacement
 
   ${cmd('nwb serve-preact-app')} ${opt('[entry]')}
     Serve a Preact app from ${opt('entry')}