+++ src/commands/build-web-module.js
@@ -6,7 +6,9 @@ import moduleBuild from '../moduleBuild'
 export default function buildModule(args, cb) {
   moduleBuild(args, {
     babel: {
-      stage: 1,
+      runtime: {
+        helpers: false
+      }
     }
   }, cb)
 }